colheader - fixed a compilation warning r=TBD
git-svn-id: http://svn.osafoundation.org/chandler/trunk@5120 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ internal/wxPython-2.5/src/generic/colheader.cpp
@@ -2334,7 +2334,7 @@ void wxColumnHeaderItem::TruncateLabelText(
 {
 wxString		truncStr;
 
-	if ((cutoffCharCount > 0) && (cutoffCharCount <= targetStr.length()))
+	if ((cutoffCharCount > 0) && (cutoffCharCount <= (long)(targetStr.length())))
 	{
 		truncStr = targetStr.Left( cutoffCharCount );
 		targetStr = truncStr + wxString( wxT("...") );
